Use datanode to assert DDL IT (#8838)
* Use datanode assert DDL IT

* Merge proxy and jdbc GeneralDDLIT

* Move assertTableMetaData to parent class

* Assert drop table
