+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-test-suite/src/test/java/org/apache/shardingsphere/test/integration/engine/it/ddl/BaseDDLIT.java
@@ -20,8 +20,13 @@ package org.apache.shardingsphere.test.integration.engine.it.ddl;
 import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
 import org.apache.shardingsphere.infra.database.type.DatabaseType;
+import org.apache.shardingsphere.infra.datanode.DataNode;
+import org.apache.shardingsphere.sharding.algorithm.sharding.inline.InlineExpressionParser;
 import org.apache.shardingsphere.test.integration.cases.assertion.ddl.DDLIntegrateTestCaseAssertion;
 import org.apache.shardingsphere.test.integration.cases.assertion.root.SQLCaseType;
+import org.apache.shardingsphere.test.integration.cases.dataset.metadata.DataSetColumn;
+import org.apache.shardingsphere.test.integration.cases.dataset.metadata.DataSetIndex;
+import org.apache.shardingsphere.test.integration.cases.dataset.metadata.DataSetMetadata;
 import org.apache.shardingsphere.test.integration.engine.it.SingleIT;
 import org.apache.shardingsphere.test.integration.env.EnvironmentPath;
 import org.apache.shardingsphere.test.integration.env.dataset.DataSetEnvironmentManager;
@@ -34,9 +39,19 @@ import org.junit.BeforeClass;
 import javax.xml.bind.JAXBException;
 import java.io.IOException;
 import java.sql.Connection;
+import java.sql.DatabaseMetaData;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.text.ParseException;
+import java.util.Collection;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertThat;
 
 public abstract class BaseDDLIT extends SingleIT {
     
@@ -89,6 +104,114 @@ public abstract class BaseDDLIT extends SingleIT {
         try (PreparedStatement preparedStatement = connection.prepareStatement(String.format("DROP TABLE %s", ((DDLIntegrateTestCaseAssertion) getAssertion()).getTable()))) {
             preparedStatement.executeUpdate();
         } catch (final SQLException ignored) {
+        }
+    }
+    
+    protected final void assertTableMetaData() throws SQLException {
+        String tableName = ((DDLIntegrateTestCaseAssertion) getAssertion()).getTable();
+        DataSetMetadata expected = getDataSet().findMetadata(tableName);
+        Collection<DataNode> dataNodes = new InlineExpressionParser(expected.getDataNodes()).splitAndEvaluate().stream().map(DataNode::new).collect(Collectors.toList());
+        if (expected.getColumns().isEmpty()) {
+            assertFalse(containsTable(dataNodes));
+            return;
+        }
+        assertTableMetaData(getActualColumns(dataNodes), getActualIndexes(dataNodes), expected);
+    }
+    
+    private void assertTableMetaData(final List<DataSetColumn> actualColumns, final List<DataSetIndex> actualIndexes, final DataSetMetadata expected) {
+        for (DataSetColumn each : expected.getColumns()) {
+            assertColumnMetaData(actualColumns, each);
+        }
+        for (DataSetIndex each : expected.getIndexes()) {
+            assertIndexMetaData(actualIndexes, each);
+        }
+    }
+    
+    private boolean containsTable(final Collection<DataNode> dataNodes) throws SQLException {
+        for (DataNode each : dataNodes) {
+            try (Connection connection = getActualDataSources().get(each.getDataSourceName()).getConnection()) {
+                if (containsTable(connection, each.getTableName())) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+    
+    private boolean containsTable(final Connection connection, final String tableName) throws SQLException {
+        return connection.getMetaData().getTables(null, null, tableName, new String[]{"TABLE"}).next();
+    }
+    
+    private List<DataSetColumn> getActualColumns(final Collection<DataNode> dataNodes) throws SQLException {
+        List<DataSetColumn> result = new LinkedList<>();
+        for (DataNode each : dataNodes) {
+            try (Connection connection = getActualDataSources().get(each.getDataSourceName()).getConnection()) {
+                result.addAll(getActualColumns(connection, each.getTableName()));
+            }
+        }
+        return result;
+    }
+    
+    private List<DataSetColumn> getActualColumns(final Connection connection, final String tableName) throws SQLException {
+        DatabaseMetaData metaData = connection.getMetaData();
+        try (ResultSet resultSet = metaData.getColumns(null, null, tableName, null)) {
+            List<DataSetColumn> result = new LinkedList<>();
+            while (resultSet.next()) {
+                DataSetColumn each = new DataSetColumn();
+                each.setName(resultSet.getString("COLUMN_NAME"));
+                each.setType(resultSet.getString("TYPE_NAME").toLowerCase());
+                result.add(each);
+            }
+            return result;
+        }
+    }
+    
+    private List<DataSetIndex> getActualIndexes(final Collection<DataNode> dataNodes) throws SQLException {
+        List<DataSetIndex> result = new LinkedList<>();
+        for (DataNode each : dataNodes) {
+            try (Connection connection = getActualDataSources().get(each.getDataSourceName()).getConnection()) {
+                result.addAll(getActualIndexes(connection, each.getTableName()));
+            }
+        }
+        return result;
+    }
+    
+    private List<DataSetIndex> getActualIndexes(final Connection connection, final String tableName) throws SQLException {
+        DatabaseMetaData metaData = connection.getMetaData();
+        try (ResultSet resultSet = metaData.getIndexInfo(null, null, tableName, false, false)) {
+            List<DataSetIndex> result = new LinkedList<>();
+            while (resultSet.next()) {
+                DataSetIndex each = new DataSetIndex();
+                each.setName(resultSet.getString("INDEX_NAME"));
+                each.setUnique(!resultSet.getBoolean("NON_UNIQUE"));
+                each.setColumns(resultSet.getString("COLUMN_NAME"));
+                result.add(each);
+            }
+            return result;
+        }
+    }
+    
+    // TODO need to assert line by line
+    private void assertColumnMetaData(final List<DataSetColumn> actual, final DataSetColumn expect) {
+        for (DataSetColumn each : actual) {
+            if (expect.getName().equals(each.getName())) {
+                if ("MySQL".equals(getDatabaseType().getName()) && "integer".equals(expect.getType())) {
+                    assertThat(each.getType(), is("int"));
+                } else if ("PostgreSQL".equals(getDatabaseType().getName()) && "integer".equals(expect.getType())) {
+                    assertThat(each.getType(), is("int4"));
+                } else {
+                    assertThat(each.getType(), is(expect.getType()));
+                }
+            }
+        }
+    }
+    
+    // TODO need to assert line by line
+    private void assertIndexMetaData(final List<DataSetIndex> actual, final DataSetIndex expect) {
+        for (DataSetIndex each : actual) {
+            if (expect.getName().equals(each.getName())) {
+                assertThat(each.isUnique(), is(expect.isUnique()));
+            }
         }
     }
 }
