+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-test-suite/src/test/resources/integrate/cases/ddl/dataset/db/create_unique_index.xml
@@ -15,4 +15,10 @@
   ~ limitations under the License.
   -->
 
-<dataset />
+<dataset>
+    <metadata table-name="t_order_details" data-nodes="db_${0..9}.t_order_details">
+        <column name="id" type="integer" />
+        <column name="description" type="varchar" />
+        <index name="t_order_details_index_t_order_details" column="id" unique="true" />
+    </metadata>
+</dataset>
