+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-test-suite/src/test/resources/integrate/cases/ddl/dataset/dbtbl_with_replica_query/alter_table.xml
@@ -16,7 +16,7 @@
   -->
 
 <dataset>
-    <metadata table-name="t_order_details">
+    <metadata table-name="t_order_details" data-nodes="primary_ds_${0..9}.t_order_details_${0..9},replica_ds_${0..9}.t_order_details_${0..9}">
         <column name="id" type="integer" />
         <column name="description" type="varchar" />
         <column name="name" type="varchar" />
