+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-test-suite/src/test/resources/integrate/cases/ddl/dataset/shadow/create_unique_index.xml
@@ -16,7 +16,7 @@
   -->
 
 <dataset>
-    <metadata table-name="t_order_details">
+    <metadata table-name="t_order_details" data-nodes="db.t_order_details">
         <column name="id" type="integer" />
         <column name="description" type="varchar" />
         <index name="t_order_details_index_t_order_details" column="id" unique="true" />
